Fix broken desktop build
